requirements.txt
streamlit
openai
google-generativeai
gspread
google-auth
pandas
python-docx
beautifulsoup4
httpx
google-search-results
pytrends
